requirements.txt
coverage
factory_boy
nose2
pytest
pytest-django